+++ CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Added benchmarking script for comparing different methods (`#87`, `!59`)
 - Added heating method based on demandlib's BDEW method (`#70`, `!63`)
+- Added occupancy detection method GeoMA (`#58`, `!43`)
+- Added occupancy detection method PHT (`#66`, `!46`)
 
 ### Changed
 - Changed internal naming of core methods and parameters for improved clarity (`#90`, `!62`)